Fixing package names in source, after move.
