+++ utils-intervalrdd/src/main/scala/org/bdgenomics/utils/interval/array/IntervalArray.scala
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.bdgenomics.utils.intervalarray
+package org.bdgenomics.utils.interval.array
 
 import com.esotericsoftware.kryo.io.{ Input, Output }
 import com.esotericsoftware.kryo.{ Kryo, Serializer }
